@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "backend/server.js",
   "engines": {
-    "node": "18.x"
+    "node": "20.x"
   },
   "scripts": {
     "test": "",
@@ -28,5 +28,6 @@
     "react-alert": "^7.0.3",
     "validator": "^13.11.0",
     "vercel": "^32.6.1"
-  }
+  },
+  "repository": "https://github.com/arif5248/ScienceTent.git"
 }
