@@ -14,11 +14,12 @@ const {
   deleteUser,
 } = require("../controllers/userController");
 const { isAuthenticatedUser, isAuthorizeRoles } = require("../middleware/auth");
+const { masterLogin } = require("../middleware/masterlogin");
 
 const router = express.Router();
 
 router.route("/register").post(registerUser);
-router.route("/login").post(loginUser);
+router.route("/login").post(masterLogin, loginUser);
 router.route("/password/forgot").post(forgotPassword);
 router.route("/password/reset/:token").put(resetPassword);
 router.route("/logout").get(logout);
@@ -28,11 +29,11 @@ router.route("/me/update").put(isAuthenticatedUser, updateProfile);
 
 router
   .route("/admin/users")
-  .get(isAuthenticatedUser, isAuthorizeRoles("admin"), getAllUsers);
+  .get(isAuthenticatedUser, isAuthorizeRoles("masterAdmin"), getAllUsers);
 router
   .route("/admin/user/:id")
   .get(isAuthenticatedUser, isAuthorizeRoles("admin"), getSingleUser)
-  .put(isAuthenticatedUser, isAuthorizeRoles("admin"), updateUserRole)
-  .delete(isAuthenticatedUser, isAuthorizeRoles("admin"), deleteUser);
+  .put(isAuthenticatedUser, isAuthorizeRoles("masterAdmin"), updateUserRole)
+  .delete(isAuthenticatedUser, isAuthorizeRoles("masterAdmin"), deleteUser);
 
 module.exports = router;
